Move files in main folder to utils
