--- commands.lua
@@ -1,2 +1,0 @@
-require 'commands.misc'
-require 'commands.where'
