+++ control.lua
@@ -5,6 +5,7 @@ _DUMP_ENV = false
 
 require 'utils.server'
 require 'utils.server_commands'
+require 'utils.command_handler'
 require 'utils.utils'
 require 'utils.pause_game'
 require 'utils.table'
@@ -18,14 +19,15 @@ require 'utils.datastore.quickbar_data'
 require 'utils.datastore.message_on_join_data'
 require 'utils.datastore.player_tag_data'
 require 'utils.datastore.supporters'
-require 'chatbot'
-require 'commands'
-require 'antigrief'
+require 'utils.chatbot'
+require 'utils.commands'
+require 'utils.antigrief'
+require 'utils.debug.command'
 require 'modules.corpse_markers'
 require 'modules.floaty_chat'
 require 'modules.show_inventory'
 require 'modules.inserter_drops_pickup'
-require 'utils.debug.command'
+require 'modules.autostash'
 
 require 'comfy_panel.main'
 require 'comfy_panel.player_list'
@@ -34,8 +36,6 @@ require 'comfy_panel.group'
 require 'comfy_panel.poll'
 require 'comfy_panel.score'
 require 'comfy_panel.config'
-
-require 'modules.autostash'
 
 ---------------- !ENABLE MODULES HERE ----------------
 --require 'modules.admins_operate_biters'
