+++ maps/fish_defender_v2/main.lua
@@ -16,7 +16,7 @@ local Server = require 'utils.server'
 local Session = require 'utils.datastore.session_data'
 local Poll = require 'comfy_panel.poll'
 local Score = require 'comfy_panel.score'
-local AntiGrief = require 'antigrief'
+local AntiGrief = require 'utils.antigrief'
 local Core = require 'utils.core'
 local format_number = require 'util'.format_number
 local random = math.random
