+++ maps/mountain_fortress_v3/main.lua
@@ -37,9 +37,9 @@ local Task = require 'utils.task'
 local Token = require 'utils.token'
 local Alert = require 'utils.alert'
 local BottomFrame = require 'comfy_panel.bottom_frame'
-local AntiGrief = require 'antigrief'
-local Misc = require 'commands.misc'
-local Modifiers = require 'player_modifiers'
+local AntiGrief = require 'utils.antigrief'
+local Misc = require 'utils.commands.misc'
+local Modifiers = require 'utils.player_modifiers'
 local BiterHealthBooster = require 'modules.biter_health_booster_v2'
 local Reset = require 'functions.soft_reset'
 local JailData = require 'utils.datastore.jail_data'
