+++ maps/mountain_fortress_v3/soft_reset.lua
@@ -1,6 +1,6 @@
 local Server = require 'utils.server'
 local Session = require 'utils.datastore.session_data'
-local Modifers = require 'player_modifiers'
+local Modifers = require 'utils.player_modifiers'
 local WPT = require 'maps.mountain_fortress_v3.table'
 
 local mapkeeper = '[color=blue]Mapkeeper:[/color]'
