+++ maps/pidgeotto/main.lua
@@ -19,7 +19,7 @@ local Poll = require 'comfy_panel.poll'
 local boss_biter = require 'maps.pidgeotto.boss_biters'
 local FDT = require 'maps.pidgeotto.table'
 local Score = require 'comfy_panel.score'
-local AntiGrief = require 'antigrief'
+local AntiGrief = require 'utils.antigrief'
 local math_random = math.random
 local insert = table.insert
 local enable_start_grace_period = true
