+++ utils/commands.lua
@@ -1,0 +1,3 @@
+require 'utils.commands.trust_system'
+require 'utils.commands.misc'
+require 'utils.commands.where'
