+++ utils/commands/misc.lua
@@ -1,5 +1,5 @@
 local Session = require 'utils.datastore.session_data'
-local Modifiers = require 'player_modifiers'
+local Modifiers = require 'utils.player_modifiers'
 local Server = require 'utils.server'
 local Color = require 'utils.color_presets'
 local Event = require 'utils.event'
@@ -239,7 +239,10 @@ commands.add_command(
         end
         if not this.creative_are_you_sure then
             this.creative_are_you_sure = true
-            player.print('[WARNING] This command will enable creative/cheat-mode for all connected players, run this command again if you really want to do this!', Color.yellow)
+            player.print(
+                '[WARNING] This command will enable creative/cheat-mode for all connected players, run this command again if you really want to do this!',
+                Color.yellow
+            )
             return
         end
         if this.creative_enabled then
